Move CA1019 to its new location.
