+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/CSharp/Microsoft.ApiDesignGuidelines.CSharp.Analyzers.csproj
@@ -130,7 +130,6 @@
     <Compile Include="CSharpMarkAttributesWithAttributeUsage.cs" />
     <Compile Include="CSharpMarkAttributesWithAttributeUsage.Fixer.cs" />
     <Compile Include="CSharpDefineAccessorsForAttributeArguments.cs" />
-    <Compile Include="CSharpDefineAccessorsForAttributeArguments.Fixer.cs" />
     <Compile Include="CSharpUsePropertiesWhereAppropriate.cs" />
     <Compile Include="CSharpUsePropertiesWhereAppropriate.Fixer.cs" />
     <Compile Include="CSharpMarkEnumsWithFlags.cs" />
