+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/Core/DefineAccessorsForAttributeArguments.Fixer.cs
@@ -1,36 +1,156 @@
-// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
+﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Editing;
 using System.Collections.Immutable;
 using System.Composition;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Analyzer.Utilities;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CodeFixes;
+using Microsoft.CodeAnalysis.Editing;
 
 namespace Microsoft.ApiDesignGuidelines.Analyzers
-{                              
-    /// <summary>
-    /// CA1019: Define accessors for attribute arguments
-    /// </summary>
-    public abstract class DefineAccessorsForAttributeArgumentsFixer : CodeFixProvider
+{
+    [ExportCodeFixProvider(LanguageNames.CSharp, LanguageNames.VisualBasic), Shared]
+    public sealed class DefineAccessorsForAttributeArgumentsFixer : CodeFixProvider
     {
-        public sealed override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(DefineAccessorsForAttributeArgumentsAnalyzer.RuleId);
+        public override ImmutableArray<string> FixableDiagnosticIds => ImmutableArray.Create(DefineAccessorsForAttributeArgumentsAnalyzer.RuleId);
 
-        public sealed override FixAllProvider GetFixAllProvider()
+        public override async Task RegisterCodeFixesAsync(CodeFixContext context)
         {
-            // See https://github.com/dotnet/roslyn/blob/master/docs/analyzers/FixAllProvider.md for more information on Fix All Providers
-            return WellKnownFixAllProviders.BatchFixer;
+            var generator = SyntaxGenerator.GetGenerator(context.Document);
+            var root = await context.Document.GetSyntaxRootAsync(context.CancellationToken).ConfigureAwait(false);
+            var node = root.FindNode(context.Span);
+
+            // We cannot have multiple overlapping diagnostics of this id.
+            var diagnostic = context.Diagnostics.Single();
+            string fixCase;
+            if (diagnostic.Properties.TryGetValue("case", out fixCase))
+            {
+                switch (fixCase)
+                {
+                    case DefineAccessorsForAttributeArgumentsAnalyzer.AddAccessorCase:
+                        var parameter = generator.GetDeclaration(node, DeclarationKind.Parameter);
+                        if (parameter != null)
+                        {
+                            context.RegisterCodeFix(new MyCodeAction(MicrosoftApiDesignGuidelinesAnalyzersResources.CreatePropertyAccessorForParameter,
+                                                         async ct => await AddAccessor(context.Document, parameter, ct).ConfigureAwait(false)),
+                                                    diagnostic);
+                        }
+                        return;
+
+                    case DefineAccessorsForAttributeArgumentsAnalyzer.MakePublicCase:
+                        var property = generator.GetDeclaration(node, DeclarationKind.Property);
+                        if (property != null)
+                        {
+                            context.RegisterCodeFix(new MyCodeAction(MicrosoftApiDesignGuidelinesAnalyzersResources.MakeGetterPublic,
+                                                             async ct => await MakePublic(context.Document, node, property, ct).ConfigureAwait(false)),
+                                                    diagnostic);
+                        }
+                        return;
+
+                    case DefineAccessorsForAttributeArgumentsAnalyzer.RemoveSetterCase:
+                        context.RegisterCodeFix(new MyCodeAction(MicrosoftApiDesignGuidelinesAnalyzersResources.MakeSetterNonPublic,
+                                                     async ct => await RemoveSetter(context.Document, node, ct).ConfigureAwait(false)),
+                                                diagnostic);
+                        return;
+
+                    default:
+                        return;
+                }
+            }
         }
 
-        public sealed override async Task RegisterCodeFixesAsync(CodeFixContext context)
-        {                              
-            // This is to get rid of warning CS1998, please remove when implementing this analyzer
-            await new Task(() => { });
-            throw new NotImplementedException();
+        private async Task<Document> AddAccessor(Document document, SyntaxNode parameter, CancellationToken cancellationToken)
+        {
+            SymbolEditor symbolEditor = SymbolEditor.Create(document);
+            var model = await document.GetSemanticModelAsync(cancellationToken).ConfigureAwait(false);
+
+            var parameterSymbol = model.GetDeclaredSymbol(parameter, cancellationToken) as IParameterSymbol;
+            if (parameterSymbol == null)
+            {
+                return document;
+            }
+
+            // Make the first character uppercase since we are generating a property.
+            var propName = char.ToUpper(parameterSymbol.Name[0]).ToString() + parameterSymbol.Name.Substring(1);
+
+            var typeSymbol = parameterSymbol.ContainingType;
+            var propertySymbol = typeSymbol.GetMembers(propName).Where(m => m.Kind == SymbolKind.Property).FirstOrDefault();
+
+            // Add a new property
+            if (propertySymbol == null)
+            {
+                await symbolEditor.EditOneDeclarationAsync(typeSymbol,
+                                                           parameter.GetLocation(), // edit the partial declaration that has this parameter symbol.
+                                                           (editor, typeDeclaration) =>
+                                                           {
+                                                               var newProperty = editor.Generator.PropertyDeclaration(propName,
+                                                                                                                      editor.Generator.TypeExpression(parameterSymbol.Type),
+                                                                                                                      Accessibility.Public,
+                                                                                                                      DeclarationModifiers.ReadOnly);
+                                                               newProperty = editor.Generator.WithGetAccessorStatements(newProperty, null);
+                                                               editor.AddMember(typeDeclaration, newProperty);
+                                                           },
+                                                           cancellationToken).ConfigureAwait(false);
+            }
+            else
+            {
+                await symbolEditor.EditOneDeclarationAsync(propertySymbol,
+                                                          (editor, propertyDeclaration) =>
+                                                          {
+                                                              editor.SetGetAccessorStatements(propertyDeclaration, null);
+                                                              editor.SetModifiers(propertyDeclaration, editor.Generator.GetModifiers(propertyDeclaration) - DeclarationModifiers.WriteOnly);
+                                                          },
+                                                          cancellationToken).ConfigureAwait(false);
+            }
+
+            return symbolEditor.GetChangedDocuments().First();
+        }
+
+        private async Task<Document> MakePublic(Document document, SyntaxNode getMethod, SyntaxNode property, CancellationToken cancellationToken)
+        {
+            // Clear the accessibility on the getter.
+            DocumentEditor editor = await DocumentEditor.CreateAsync(document, cancellationToken).ConfigureAwait(false);
+            editor.SetAccessibility(getMethod, Accessibility.NotApplicable);
+
+            // If the containing property is not public, make it so
+            var propertyAccessibility = editor.Generator.GetAccessibility(property);
+            if (propertyAccessibility != Accessibility.Public)
+            {
+                editor.SetAccessibility(property, Accessibility.Public);
+
+                // Having just made the property public, if it has a setter with no Accessibility set, then we've just made the setter public. 
+                // Instead restore the setter's original accessibility so that we don't fire a violation with the generated code.
+                var setter = editor.Generator.GetAccessor(property, DeclarationKind.SetAccessor);
+                if (setter != null)
+                {
+                    var setterAccessibility = editor.Generator.GetAccessibility(setter);
+                    if (setterAccessibility == Accessibility.NotApplicable)
+                    {
+                        editor.SetAccessibility(setter, propertyAccessibility);
+                    }
+                }
+            }
+
+            return editor.GetChangedDocument();
+        }
+
+        private async Task<Document> RemoveSetter(Document document, SyntaxNode setMethod, CancellationToken cancellationToken)
+        {
+            DocumentEditor editor = await DocumentEditor.CreateAsync(document, cancellationToken).ConfigureAwait(false);
+            editor.SetAccessibility(setMethod, Accessibility.Internal);
+            return editor.GetChangedDocument();
+        }
+
+        private class MyCodeAction : DocumentChangeAction
+        {
+            public MyCodeAction(string title, Func<CancellationToken, Task<Document>> createChangedDocument)
+                : base(title, createChangedDocument)
+            {
+            }
         }
     }
 }
