+++ PortFxCop/src/Microsoft.ApiDesignGuidelines.Analyzers/Core/MicrosoftApiDesignGuidelinesAnalyzersResources.resx
@@ -499,4 +499,13 @@ Names of parameters and members are better used to communicate their meaning tha
   <data name="IdentifiersShouldNotContainTypeNamesMessage" xml:space="preserve">
     <value>Identifier '{0}' contains type name</value>
   </data>
+  <data name="CreatePropertyAccessorForParameter" xml:space="preserve">
+    <value>Create a property accessor.</value>
+  </data>
+  <data name="MakeGetterPublic" xml:space="preserve">
+    <value>Make the getter of the property public</value>
+  </data>
+  <data name="MakeSetterNonPublic" xml:space="preserve">
+    <value>Make the setter of the property non-public</value>
+  </data>
 </root>
